Changed the type of variables x and y from double to int.

+++ src/Characters/Vector2.java
@@ -1,10 +1,10 @@
 package Characters;
 
 public class Vector2 {
-    public float x;
-    public float y;
+    public int x;
+    public int y;
 
-    public Vector2(float x, float y) {
+    public Vector2(int x, int y) {
         this.x = x;
         this.y = y;
     }
